requirements.txt
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
ffmpeg-python==0.2.0
tqdm==4.66.1
Pillow==10.1.0